+++ cmd/stack/destroy.go
@@ -17,13 +17,10 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/okteto/okteto/cmd/utils"
 	"github.com/okteto/okteto/pkg/analytics"
-	"github.com/okteto/okteto/pkg/helm"
+	"github.com/okteto/okteto/pkg/cmd/stack"
 	"github.com/okteto/okteto/pkg/log"
 	"github.com/spf13/cobra"
-	"helm.sh/helm/v3/pkg/action"
-	"helm.sh/helm/v3/pkg/cli"
 )
 
 //Destroy destroys a stack
@@ -34,7 +31,7 @@ func Destroy(ctx context.Context) *cobra.Command {
 		Short: fmt.Sprintf("Destroys a stack"),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			name := args[0]
-			err := executeDestroyStack(ctx, name, namespace)
+			err := stack.Destroy(ctx, name, namespace)
 			analytics.TrackDestroyStack(err == nil)
 			if err == nil {
 				log.Success("Successfully destroyed stack '%s'", name)
@@ -50,37 +47,4 @@ func Destroy(ctx context.Context) *cobra.Command {
 	}
 	cmd.Flags().StringVarP(&namespace, "namespace", "n", "", "overwrites the stack namespace where the stack is destroyed")
 	return cmd
-}
-
-func executeDestroyStack(ctx context.Context, name, namespace string) error {
-	spinner := utils.NewSpinner(fmt.Sprintf("Destroying stack '%s'...", name))
-	spinner.Start()
-	defer spinner.Stop()
-
-	settings := cli.New()
-	actionConfig := new(action.Configuration)
-	if namespace == "" {
-		namespace = settings.Namespace()
-	}
-
-	if err := actionConfig.Init(settings.RESTClientGetter(), namespace, helm.HelmDriver, func(format string, v ...interface{}) {
-		message := fmt.Sprintf(format, v...)
-		spinner.Update(fmt.Sprintf("%s...", message))
-	}); err != nil {
-		return fmt.Errorf("error initializing stack client: %s", err)
-	}
-
-	exists, err := helm.ExistRelease(action.NewList(actionConfig), name)
-	if err != nil {
-		return fmt.Errorf("error listing stacks: %s", err)
-	}
-	if !exists {
-		return fmt.Errorf("stack %s does not exist", name)
-	}
-
-	uClient := action.NewUninstall(actionConfig)
-	if _, err := uClient.Run(name); err != nil {
-		return fmt.Errorf("error destroying stack '%s': %s", name, err)
-	}
-	return nil
 }
