+++ cmd/stack/stack.go
@@ -20,17 +20,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-const (
-	//HelmRepoURL link to okteto stack helm repo
-	HelmRepoURL = "https://apps.okteto.com"
-	//HelmRepoName repo name for local link to okteto stack helm repo
-	HelmRepoName = "okteto-charts"
-	//HelmChartName chart name for okteto stack
-	HelmChartName = "stacks"
-	//HelmChartVersion chart version for okteto stack chart
-	HelmChartVersion = "0.1.0"
-)
-
 //Stack stack management commands
 func Stack(ctx context.Context) *cobra.Command {
 	cmd := &cobra.Command{
