+++ pkg/analytics/analytics.go
@@ -43,8 +43,8 @@ const (
 	statusEvent          = "Status"
 	doctorEvent          = "Doctor"
 	buildEvent           = "Build"
-	deployStackEvent     = "DeployStack"
-	destroyStackEvent    = "DestroyStack"
+	deployStackEvent     = "Deploy Stack"
+	destroyStackEvent    = "Destroy Stack"
 	loginEvent           = "Login"
 	initEvent            = "Create Manifest"
 	namespaceEvent       = "Namespace"
